# part1/src/main/java/com/example/hqadministrationapi/dto/DepartmentRequest.java
package com.example.hqadministrationapi.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

import java.math.BigDecimal;

public record DepartmentRequest(
        @NotBlank String name,
        @NotNull @Positive BigDecimal budget,
        @NotBlank String location
) {}